+++ src/main/java/api/m2/movements/records/movements/MovementRecord.java
@@ -22,6 +22,7 @@ public record MovementRecord(Long id,
                              String type,
                              Integer cuotaActual,
                              Integer cuotasTotales,
+                             LocalDate lastCreditPayment,
                              Metadata metadata) {
 
     public record Metadata(UserBaseRecord owner,