+++ src/main/java/api/m2/movements/services/income/IncomeAddService.java
@@ -125,6 +125,7 @@ public class IncomeAddService {
                 currencySymbol,
                 null,
                 null,
-                bankDescription);
+                bankDescription,
+                null);
     }
 }