+++ src/main/java/api/m2/movements/services/movements/files/ExpenseFileStrategy.java
@@ -56,7 +56,8 @@ public abstract class ExpenseFileStrategy {
                 e.currency().getSymbol(),
                 cuotaActual,
                 cuotaTotales,
-                getBank()
+                getBank(),
+                null
         );
 
     }